Refactor: Replace fetch with direct function calls for approving, rejecting, and updating fallen profiles
